refactor productBlock to function component and pull sub components into new file
